backend/requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.1.2
django-cors-headers==3.5.0
django-extensions==3.0.9
django-rest-framework==0.1.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
idna==2.10
ipdb==0.13.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.7.1
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
